Extracted GetHttpRequestBody to goplugin::Util Also changed ReadJSON method to not consume request body. Duplicated buffer will be stored under request to allow multiple reads from it
